refactor(API): Change response from API to client

+++ Application/Features/Commands/DeleteTask/DeleteTaskCommandHandler.cs
@@ -1,9 +1,10 @@
-﻿using Domain.Interfaces;
+﻿using Application.Common.Exceptions;
+using Domain.Interfaces;
 using MediatR;
 
 namespace Application.Features.Commands.DeleteTask
 {
-    public class DeleteTaskCommandHandler : IRequestHandler<DeleteTaskCommand, string>
+    public class DeleteTaskCommandHandler : IRequestHandler<DeleteTaskCommand>
     {
         private readonly ITaskRepository _taskRepository;
 
@@ -12,11 +13,14 @@ namespace Application.Features.Commands.DeleteTask
             _taskRepository = taskRepository;
         }
 
-        public async Task<string> Handle(DeleteTaskCommand request, CancellationToken cancellationToken)
+        public async Task Handle(DeleteTaskCommand request, CancellationToken cancellationToken)
         {
             var task = await _taskRepository.GetTaskByIdAsync(request.Id);
+            if (task is null)
+            {
+                throw new NotFoundException("Task not found.");
+            }
             _taskRepository.DeleteTask(task);
-            return "";
         }
     }
 }
